Refactor team structure in TeamExpansionForm2023.jsx
Updated the structure of the teams in the TeamExpansionForm2023.jsx file. The change was necessary to simplify the form, improve user experience and align it with the changes in the team hierarchy. We have also renamed "interestedVerticals" with more appropriate name "interestedTeams". Specific subverticals under 'Branding' were removed, instead more generalized team categories have been added.