Update image source and enhance style of startup junction
This commit updates the source URL of the Image from .png to .jpg in the startup junction component, to ensure correct asset is loaded. Additionally, the 'mix-blend-mode' style property is added to .vcimg class which will offer a better visual integration of this element with the surrounding content. Lastly, this commit includes updates to the yarn.lock file for ensuring correct package versioning and dependencies.