Add link to Startup News on Startup Junction page
A new external link was added to the Startup Junction page. This link directs users to a site called Startup News (https://startupnews.fyi/). The purpose of this change is to provide users with easy, one-click access to additional content and resources that may be relevant to their interests in the startup scene.